enter the asset registry
