+++ CMakeLists.txt
@@ -81,6 +81,7 @@ file(GLOB_RECURSE NOTHOFAGUS_PUBLIC_HEADERS "include/*.h")
 add_library(nothofagus
     "source/canvas.cpp"
     "source/canvas_impl.cpp"
+    "source/asset_registry.cpp"
     "source/imgui_rtt_manager.cpp"
     "source/imgui_font_manager.cpp"
     ${NOTHOFAGUS_BACKEND_SRC}
